BWS-PKG - Bug 34868: Add ability for SIP2 to distinguish missing item from other lost types
The SIP circulation status specifies that a 12 means an item is lost, and 13 means an item is missing. In Koha, missing items are simply a type of lost item so we never send a 13. This is an important distinction for some SIP based inventory tools. It would be good to be able to specify when lost status means "missing" at the SIP login level.

Test Plan:
1) Apply this patch
2) prove t/db_dependent/SIP/Transaction.t

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 34868: (QA follow-up) Rename new option, add comments

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 34868: (QA follow-up) Fix XML

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Bug 34868: (QA follow-up) Tidy sip_circulation_status

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>